--- a/pkg/rpc/settings.go
+++ b/pkg/rpc/settings.go
@@ -119,7 +119,8 @@ var sourceAddr = func() net.Addr {
 }()
 
 type serverOpts struct {
-	interceptor func(fullMethod string) error
+	interceptor        func(fullMethod string) error
+	metricsInterceptor RequestMetricsInterceptor
 }
 
 // ServerOption is a configuration option passed to NewServer.
@@ -140,5 +141,12 @@ func WithInterceptor(f func(fullMethod string) error) ServerOption {
 				return f(fullMethod)
 			}
 		}
+	}
+}
+
+// WithMetricsServerInterceptor adds a RequestMetricsInterceptor to the grpc server.
+func WithMetricsServerInterceptor(interceptor RequestMetricsInterceptor) ServerOption {
+	return func(opts *serverOpts) {
+		opts.metricsInterceptor = interceptor
 	}
 }