… docs/generated/metrics/metrics.html
… pkg/roachprod/opentelemetry/cockroachdb_metrics.go