Adds a new grpc.UnaryServerInterceptor that records the duration of a grpc request. The metric is labeled by the grpc full method name and status code. This is disabled by default and can be enabled by setting `server.grpc.metrics.enabled` to true.

This metric uses metric.HistogramVec and is exported to external collectors.

Epic: None
Release note: None