@kyle-a-wong 2025-02-25
Ran the following micro bench mark with this `server.grpc.metrics.enabled` set to true:
`benchdiff ./pkg/sql/tests -b -r 'BenchmarkSysbench/SQL/3node/oltp_read_write' --benchtime 10x`
and got the following results:
```
name                                   old time/op    new time/op    delta
Sysbench/SQL/3node/oltp_read_write-12    6.69ms ± 5%    6.57ms ± 4%    ~     (p=0.095 n=10+9)

name                                   old errs/op    new errs/op    delta
Sysbench/SQL/3node/oltp_read_write-12      0.00           0.00         ~     (all equal)

name                                   old alloc/op   new alloc/op   delta
Sysbench/SQL/3node/oltp_read_write-12    2.38MB ± 2%    2.36MB ± 0%  -1.00%  (p=0.012 n=10+8)

name                                   old allocs/op  new allocs/op  delta
Sysbench/SQL/3node/oltp_read_write-12     13.2k ± 2%     13.1k ± 1%    ~     (p=0.481 n=10+10)
```